requirements.txt
django==4.1.2
psycopg2==2.9.4
django-environ==0.9.0
djangorestframework==3.14.0
django-heroku==0.3.1
gunicorn==20.1.0
pyyaml==6.0
uritemplate==4.1.1
